backend/pyproject.toml
[tool.poetry]
name = "Template project"
version = "0.1.0"
description = "template project"
license = "MIT"
authors=["Egor Ponyavin <ep@idaproject.com>"]

[tool.poetry.dependencies]
# core
python = "3.11"
Django = "4.1.*"
certifi = "2022.9.24"
psycopg2-binary = "2.9.*"
sentry-sdk = "1.5.*"
requests = "2.26.*"
boto3 = "1.20.*"
django-redis-cache = "3.0.*"
django-storages = "1.12.*"
gunicorn = "^20.1.0"
django-user-agents = "^0.4.0"
django-hashedfilenamestorage = {git = "https://github.com/ecometrica/django-hashedfilenamestorage.git", rev = "45b7ef1e68b0e89c2fd6a20d2d1db57f620a5f8f"}
python-telegram-bot = "^20.3"
pyTelegramBotAPI = "^4.12.0"

# Models \ Admin
django-solo = "2.0.*"
django-ckeditor = "6.2.*"
django-admin-sortable2 = "2.1.*"
django-location-field = "~2.1.0"
phonenumbers = "^8.12.29"
django-phonenumber-field = "5.2.*"
django-admin-list-filter-dropdown = "~1.0.3"

# Data
xmltodict = "0.13.0"

# Rest
djangorestframework = "3.14.*"
django-filter = "22.*"
drf-spectacular = "0.24.*"

# Test
factory-boy = "^3.2.0"
pytest = "^6.2.4"
pytest-mock = "^3.6.1"
pytest-cov = "^2.12.1"
pytest-django = "^4.4.0"
pytest-sugar = "^0.9.4"

# Celery
celery = "^5.1.2"
django-celery-email = "~3.0.0"

# Others
Pillow = "9.3.*"
django-robots = "5.0.*"
beautifulsoup4 = "4.12.*"
python-slugify = "^6.1.2"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.1"

# Linting
flake8 = "6.0.*"
flake8-alphabetize = "0.0.*"
flake8-bandit = "4.1.*"
flake8-black = "0.3.*"
flake8-bugbear = "*"
flake8-builtins = "2.1.*"
flake8-class-attributes-order = "0.1.*"
flake8-comprehensions = "3.10.*"
flake8-debugger = "4.1.*"
flake8-docstrings = "1.7.*"
flake8-isort = "6.0.*"
flake8-pie = "0.16.*"
flake8-print = "5.0.*"
flake8-printf-formatting = "1.1.*"
flake8-pytest-style = "1.7.*"
flake8-walrus = "1.2.*"

# Formatting
black = "23.1.*"
isort = "5.12.*"

# Typehinting
mypy = "1.*"
mypy-extensions = "1.*"
django-stubs = "1.14.*"
djangorestframework-stubs="1.8.*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
multi_line_output = 3
line_length = 100
include_trailing_comma = true
